refactor(routes): extract /bugs to its own module — god-file decomposition pilot
First domain out of the 6852-line god file via the safe, verified process:
- /bugs (6 routes) -> server/routes/bugs.js; handlers VERBATIM, helpers via deps,
  DB fns imported from db.js. mycelium.js 6852 -> 6757.
- manifest hardened to capture middleware COUNT (a dropped agentWriteLimiter now fails).

Four gates, all green:
  1. route-manifest --check unchanged (284 routes)
  2. full suite green (260 tests)
  3. move-only diff (103 del, +8 import/register)
  4. handler bodies byte-identical to git HEAD, same order (non-Claude byte-truth)

The reference cycle the squad now replicates across the remaining ~41 domains.
